refactor: session questionnaire
