+++ src/screens/etis/sessionQuestionnaire/components/ResultViews.tsx
